backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
google-genai>=1.0.0
httpx>=0.27.0
supabase>=2.10.0
python-dotenv>=1.0.1
razorpay>=1.4.1
python-multipart>=0.0.12
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
catboost>=1.2.7
joblib>=1.4.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0